% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ls.R
\name{opt_ls}
\alias{opt_ls}
\title{LS through optimization}
\usage{
opt_ls(y, X)
}
\arguments{
\item{y}{A \code{vector} of response.}

\item{X}{A \code{matrix} of predictors.}
}
\value{
A \code{list} containing the following attributes:
\describe{XX}
}
\description{
XXX
}
\examples{
n <- 100
set.seed(123)
X <- cbind(rnorm(n), rpois(n, lambda = 3))
beta <- c(2, -3, 1) # true beta
design <- cbind(rep(1, n), X)
y <- design \%*\% beta + rnorm(n)
opt_ls(y = y, X = X)
}
\author{
Roberto Molinari
}
